Move BitcoinNetwork enum to threshold lib
