--- src/types/bitcoin.ts
@@ -1,4 +1,0 @@
-export enum BitcoinNetwork {
-  mainnet = "mainnet",
-  testnet = "testnet",
-}
