+++ src/types/index.ts
@@ -8,7 +8,6 @@ export * from "./page"
 export * from "./rewards"
 export * from "./staking"
 export * from "./staking-applications"
-export * from "./bitcoin"
 
 export type FetchingState<DataType> = {
   isInitialFetchDone?: boolean
